./src/ANL/ANL_CC1npip_Evt_1DcosmuStar_nu.h

+++ src/FitBase/BaseFitEvt.h
@@ -186,7 +186,7 @@ class BaseFitEvt {
 
   // New Splines
   //SplineReader* fSplineReader;
-  double* fSplineCoeff;
+  float* fSplineCoeff;
   SplineReader* fSplineRead;
   double SavedRWWeight;
   // double fDialCoeffs;
